# src/appconfigs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Welcome Agri-Vesty API",
    "description": "Agri-vesty is a crowdfunding Platform. The aim of Agri-vesty is to help farmers to raise farming capital. In this platform, a user can either create a crowdfunding farming project or invest money towards their favorite farming projects.",
    "contact": {
      "name": "Team-179-Group-A",
      "url": "https://agri-vesty.herokuapp.com/"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:{port}/{basePath}",
      "description": "Local server",
      "variables": {
        "port": {
          "enum": [
            "8443",
            "443"
          ],
          "default": "4000"
        },
        "basePath": {
          "default": "api"
        }
      }
    },
    {
      "url": "https://agri-vesty.herokuapp.com/{basePath}",
      "description": "Production server",
      "variables": {
        "basePath": {
          "default": "api"
        }
      }
    }
  ],
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "components": {
    "schemas": {
      "GeneralError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "projects": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "dateStarted": {
            "type": "string",
            "format": "date-time"
          },
          "dateEnded": {
            "type": "string",
            "format": "date-time"
          },
          "targetCost": {
            "type": "number",
            "format": "double"
          },
          "totalInvested": {
            "type": "number",
            "format": "double"
          },
          "active": {
            "type": "boolean"
          },
          "isWithdrawable": {
            "type": "boolean"
          }
        }
      },
...
    },
...
    "responses": {
      "Error": {
        "description": "General Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GeneralError"
            }
          }
        }
      },
...
    },
    "securitySchemes":{
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "Bearer": {
        "description": "JWT Token",
        "name": "Authorization",
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/project/{id}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Updates a  single project",
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "create the title of the project",
                    "type": "string"
                  },
                  "description": {
                    "description": "create the description of your project",
                    "type": "string"
                  },
                  "imageUrl": {
                    "description": "image of your project",
                    "type": "string",
                    "format": "binary"
                  },
                  "targetCost": {
                    "description": "The target amount",
                    "type": "number",
                    "format": "double"
                  },
                  "dateStarted": {
                    "description": "Date you started your project",
                    "type": "string",
                    "format": "date-time"
                  },
                  "dateEnded": {
                    "description": "The date your project will end",
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "imageUrl",
                  "targetCost",
                  "dateStarted",
                  "dateEnded"
                ]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "create the title of the project",
                    "type": "string"
                  },
                  "description": {
                    "description": "create the description of your project",
                    "type": "string"
                  },
                  "imageUrl": {
                    "description": "image of your project",
                    "type": "string"
                  },
                  "targetCost": {
                    "description": "The target amount",
                    "type": "number",
                    "format": "double"
                  },
                  "dateStarted": {
                    "description": "Date you started your project",
                    "type": "string",
                    "format": "date-time"
                  },
                  "dateEnded": {
                    "description": "The date your project will end",
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "imageUrl",
                  "targetCost",
                  "dateStarted",
                  "dateEnded"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated a single project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/projects"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                },
                "example": {
                  "message": "Method not allowed",
                  "internal_code": "405"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                },
                "example": {
                  "message": "Not Found",
                  "internal_code": "404"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/project/{projectid}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a single project",
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a single project.",
            "content": {
              "application/json": {
                "schema": {
                  "type":"string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                },
                "example": {
                  "message": "Method not allowed",
                  "internal_code": "405"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/Error"
                },
                "example": {
                  "message": "Not Found",
                  "internal_code": "404"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  }
}